add the champion player that supose to win the dummy players

+++ Board.cpp
@@ -4,7 +4,7 @@
 
 #include "Board.h"
 
-Board::Board(size_t a) :board(new square[a*a]), length(a){ //new constracrot
+Board::Board(uint a) :board(new square[a*a]), length(a){ //new constracrot
 
 }
